moved swmm interface files to file i/o
